+++ src/power_pptx/__init__.py
@@ -32,7 +32,7 @@ from power_pptx.parts.slide import (
 if TYPE_CHECKING:
     from power_pptx.opc.package import Part
 
-__version__ = "2.1.1"
+__version__ = "2.2.0"
 
 sys.modules["power_pptx.exceptions"] = exceptions
 del sys